src/isar/broker/subscriber/requirements.txt
paho-mqtt
requests
black
PyYAML
